==> 01_Data_Preparation/02_Data_Wrangling.ipynb <==
import pandas as pd
import numpy as np
avaliacao = pd.read_csv('dados/avaliacao.csv')
produtor = pd.read_csv('dados/produtor_crm.csv')
movimento = pd.read_csv('dados/movimento.csv')

import pandas_profiling
movimento[movimento.cod_produtor==12]

import matplotlib.pyplot as plt

def get_corr(df):

    f = plt.figure(figsize=(24, 24))
    plt.matshow(df.corr(), fignum=f.number)
    plt.xticks(range(df.shape[1]), df.columns, fontsize=12, rotation=90)
    plt.yticks(range(df.shape[1]), df.columns, fontsize=12)
    cb = plt.colorbar()
    cb.ax.tick_params(labelsize=16)
    plt.plot()
    #plt.title('Correlation Matrix', fontsize=16);
    
get_corr(pd.read_csv('dados/avaliacao.csv'))

df_movimento = pd.read_csv('./dados/movimento.csv')
df_safra = df_movimento[df_movimento.Codcult=='SOT']

df_1 = df_safra.loc[((df_safra.safra=='2014/2015') & (df_safra.vlr_total_fat>0))]
df_2 = df_safra.loc[((df_safra.safra=='2015/2016') & (df_safra.vlr_total_fat>0))]
df_3 = df_safra.loc[((df_safra.safra=='2016/2017') & (df_safra.vlr_total_fat>0))]
df_4 = df_safra.loc[((df_safra.safra=='2018/2019') & (df_safra.vlr_total_fat>0))]
df_5 = df_safra.loc[((df_safra.safra=='2016/2017') & (df_safra.vlr_total_fat==0))]
df_6 = df_safra.loc[((df_safra.safra=='2018/2019') & (df_safra.vlr_total_fat==0))]

resultado = pd.DataFrame()

for id_ in range(1000):
    if (df_1.vlr_total_fat.loc[((df_1.cod_produtor == id_))].all() > 0):
        if (df_2.vlr_total_fat.loc[((df_2.cod_produtor == id_))].all() > 0):
            if (df_3.vlr_total_fat.loc[((df_3.cod_produtor == id_))].all() > 0):
                if (df_4.vlr_total_fat.loc[((df_4.cod_produtor == id_))].all() > 0):
                    df1 = df_safra.loc[((df_safra.safra=='2014/2015') & (df_safra.vlr_total_fat>0) & (df_safra.cod_produtor == id_))]
                    df2 = df_safra.loc[((df_safra.safra=='2015/2016') & (df_safra.vlr_total_fat>0) & (df_safra.cod_produtor == id_))]
                    df3 = df_safra.loc[((df_safra.safra=='2016/2017') & (df_safra.vlr_total_fat>0) & (df_safra.cod_produtor == id_))]
                    df4 = df_safra.loc[((df_safra.safra=='2018/2019') & (df_safra.vlr_total_fat>0) & (df_safra.cod_produtor == id_))]
                    frames = [df1, df2, df3, df4]
                    resultado = resultado.append(frames)

resultado

resultado2 = pd.DataFrame()

for id_ in range(1000):
    if (df_1.vlr_total_fat.loc[((df_1.cod_produtor == id_))].all() > 0):
        if (df_2.vlr_total_fat.loc[((df_2.cod_produtor == id_))].all() > 0):
            if (df_5.vlr_total_fat.loc[((df_5.cod_produtor == id_))].all() == 0):
                if (df_6.vlr_total_fat.loc[((df_6.cod_produtor == id_))].all() == 0):
                    df1 = df_safra.loc[((df_safra.safra=='2014/2015') & (df_safra.vlr_total_fat>0) & (df_safra.cod_produtor == id_))]
                    df2 = df_safra.loc[((df_safra.safra=='2015/2016') & (df_safra.vlr_total_fat>0) & (df_safra.cod_produtor == id_))]
                    df5 = df_safra.loc[((df_safra.safra=='2016/2017') & (df_safra.vlr_total_fat==0) & (df_safra.cod_produtor == id_))]
                    df6 = df_safra.loc[((df_safra.safra=='2018/2019') & (df_safra.vlr_total_fat==0) & (df_safra.cod_produtor == id_))]
                    frames = [df1, df2, df5, df6]
                    resultado2 = resultado2.append(frames)

resultado2

#modelo criado a partir do faturamento
#Classificador do cooperado 0: nao vai ficar inativo, 1 fica inativo, 2 alerta de inatividade
# calcular com base na safra em quantas safras o produtor vai comecar a ficar inativo
# dropar as colunas 2016/2017 -> 2018/2019 para todos os produtores para o modelo não usar os ZEROS de faturamento.
#2013/2014 -> 2016/2017 -> 6 safras
#2014/2015 -> 5 safras
#produtor que nao fica inativo mais que 7 safras
# cria-se 3 classificações produtor fica inativo em 2 safras, produtor fica inativo entre 3 a 5 safras > 6 safras


resultado['classificador'] = 0
resultado2['classificador'] = 1

avaliacao = pd.read_csv('dados/avaliacao.csv')
avaliacao.rename(columns={'nome-abrev':'cod_nome'}, inplace=True)

res_ids = resultado.cod_produtor.unique()
res_ids2 = resultado2.cod_produtor.unique()

def busca_dados(id_, df_1, dat_):
    if (df_1.loc[df_1.cod_produtor == id_].any()).any():
        #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_1.vlr_total_fat.loc[(df_1.safra==dat_)& (df_1.cod_produtor==id_)].sum()
    
for i in range(1000):
    
    avaliacao.loc[avaliacao.cod_nome==i, '2014/2014'] = busca_dados(i, df_movimento,'2014/2014')

def busca_dados(id_, df_1,df_2):
    if (df_1.loc[df_1.cod_produtor == id_].any()).any():
        #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum()
    if (df_2.loc[df_2.cod_produtor == id_].any()).any():
    #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_2.vlr_total_fat.loc[(df_2.safra=='2014/2015')& (df_2.cod_produtor==id_)].sum()
    
for i in range(1000):
    avaliacao.loc[avaliacao.cod_nome==i, '2014/2015'] = busca_dados(i, resultado,resultado2)

def busca_dados(id_, df_1, dat_):
    if (df_1.loc[df_1.cod_produtor == id_].any()).any():
        #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_1.vlr_total_fat.loc[(df_1.safra==dat_)& (df_1.cod_produtor==id_)].sum()
    
for i in range(1000):
    
    avaliacao.loc[avaliacao.cod_nome==i, '2015/2015'] = busca_dados(i, df_movimento,'2015/2015')

df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015') & (df_1.cod_produtor==5)]


def busca_dados(id_, df_1,df_2, dat_):
    if (df_1.loc[df_1.cod_produtor == id_].any()).any():
        #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_1.vlr_total_fat.loc[(df_1.safra==dat_)& (df_1.cod_produtor==id_)].sum()
    if (df_2.loc[df_2.cod_produtor == id_].any()).any():
    #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_2.vlr_total_fat.loc[(df_2.safra==dat_)& (df_2.cod_produtor==id_)].sum()

for i in range(1000):
    
    avaliacao.loc[avaliacao.cod_nome==i, '2015/2016'] = busca_dados(i, resultado,resultado2,'2015/2016')

def busca_dados(id_, df_1, dat_):
    if (df_1.loc[df_1.cod_produtor == id_].any()).any():
        #print(df_1.vlr_total_fat.loc[(df_1.safra=='2014/2015')& (df_1.cod_produtor==id_)].sum())
        return df_1.vlr_total_fat.loc[(df_1.safra==dat_)& (df_1.cod_produtor==id_)].sum()
    
for i in range(1000):
    
    avaliacao.loc[avaliacao.cod_nome==i, '2016/2016'] = busca_dados(i, df_movimento,'2016/2016')

df_movimento.safra.unique()

avaliacao

avaliacao.dropna(inplace=True)

def busca_classificador(id_, res):

        return res.classificador.loc[res.cod_produtor==id_].mean()
    
for i in res_ids:
    avaliacao.loc[avaliacao.cod_nome==i, 'classificador'] = busca_classificador(i, resultado)
for i in res_ids2:
    avaliacao.loc[avaliacao.cod_nome==i, 'classificador'] = busca_classificador(i, resultado2)
#avaliacao['classificador'] = sum(avaliacao.classificador1 + avaliacao.classificador2 + avaliacao.classificador3)
#avaliacao.drop(['classificador1','classificador2','classificador3'], axis=1, inplace=True)

avaliacao[avaliacao.classificador==1]

avaliacao['soma_fat']= avaliacao['2014/2014'] + avaliacao['2014/2015'] + avaliacao['2015/2015'] + avaliacao['2015/2016'] + avaliacao['2016/2016']

index_values = avaliacao.loc[avaliacao.soma_fat<5000].index

avaliacao.drop(index_values , inplace=True)

avaliacao

avaliacao[avaliacao.classificador==1]
#avaliacao.drop(['Unnamed: 0'], axis=1, inplace=True)
#avaliacao.drop(['cod_empresa'], axis=1, inplace=True)
avaliacao.drop(['cod_estabel', 'Bloqueado', 'log_ativo','dt_ult_movto','safra_ult_movto','situacao','cdn_cliente'], axis=1, inplace=True)

avaliacao['dt_inscricao'] = pd.to_datetime(avaliacao.dt_inscricao)

avaliacao['cooperado_faz_x_anos'] = 2020 - avaliacao.dt_inscricao.dt.year

avaliacao.drop('dt_inscricao', axis=1, inplace=True)

avaliacao.columns

import matplotlib.pyplot as plt

def get_corr(df):

    f = plt.figure(figsize=(24, 24))
    plt.matshow(df.corr(), fignum=f.number)
    plt.xticks(range(df.shape[1]), df.columns, fontsize=12, rotation=90)
    plt.yticks(range(df.shape[1]), df.columns, fontsize=12)
    cb = plt.colorbar()
    cb.ax.tick_params(labelsize=16)
    plt.plot()
    #plt.title('Correlation Matrix', fontsize=16);
    
get_corr(avaliacao)

def scale(df):
    #return (df - df.mean())/df.std()
    return (df - df.min())/(df.max()-df.min())

for col in avaliacao.columns:
    if col[0] == '2':
        avaliacao[col] = scale(avaliacao[col])
    if col == 'cooperado_faz_x_anos':
        avaliacao[col] = scale(avaliacao[col])

y = pd.DataFrame(avaliacao['classificador'])
X = avaliacao.drop(['classificador', 'cod_nome'], axis=1)
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.25, random_state=4)

X_train

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_selection import SelectFromModel

sel = SelectFromModel(RandomForestClassifier(n_estimators = 100))
sel.fit(X_train, y_train)

sel.get_support()

selected_feat= X_train.columns[(sel.get_support())]
len(selected_feat)

print(selected_feat)

X_test = X_test[['ind-protesto','cooperado_faz_x_anos','tipo_retencao_inss','GRANDE2', 'GRANDE3','PRMAX','2014/2014', '2014/2015', '2015/2015', '2015/2016', '2016/2016']]

X_train = X_train[['ind-protesto','cooperado_faz_x_anos','tipo_retencao_inss','GRANDE2', 'GRANDE3','PRMAX','2014/2014', '2014/2015', '2015/2015', '2015/2016', '2016/2016']]
#X_train

from sklearn.neighbors import KNeighborsClassifier
knn = KNeighborsClassifier(n_neighbors=5)
knn.fit(X_train, y_train)
prediction = knn.predict(X_test)

from sklearn.metrics import classification_report

print(classification_report(y_test, prediction))
